i2c: Fix 32 byte SMBus transfers
In 01c3e8417, we attempted to fix 32 byte transfers (as the kernel now
properly enforced transfer size).

Unfortunately the kernel uses a 33 byte I/O buffer (as it adds a length),
so SMbus operations were corrupting the stack. Add a new symbol for the
actual transfer length (SMBusMax, 32 bytes) but make the I/O buffers 34 bytes.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>